Add files via upload
This program reads data from a .txt file that reads the scholarly journal names included and their impact factors - how many times they have been quoted in other scholarly journals divided by the number of articles in said journal. 

This serves as a template program for extracting data from a file that can be passed through a Python script in a command-line interface. For example, it can be modified for periodic web scraping of academic publication websites for gathering relevant research relating to the user's research topic for personal study or storage in a database.